chore(webpack): Move getEntityLink to /common folder
Importing server-side code in the client-side code creates some issues with Webpack, and shouldn't be done anyway.
Moving common utility functions to the src/common folder so it can be imported into src/client as well as src/server.
