+++ src/client/entity-editor/relationship-editor/relationship.tsx
@@ -23,7 +23,7 @@ import {OverlayTrigger, Tooltip} from 'react-bootstrap';
 import Entity from '../common/entity';
 import RelationshipAttribute from './relationship-attribute';
 import _ from 'lodash';
-import {getEntityLink} from '../../../server/helpers/utils';
+import {getEntityLink} from '../../../common/helpers/utils';
 
 
 function getEntityObjectForDisplay(entity: _Entity, makeLink: boolean) {
