+++ src/common/helpers/utils.ts
@@ -93,6 +93,17 @@ export function sortRelationshipOrdinal(sortByProperty: string) {
 }
 
 
+/**
+ * Returns an API path for interacting with the given Bookshelf entity model
+ *
+ * @param {object} entity - Entity object
+ * @returns {string} - URL path to interact with entity
+ */
+export function getEntityLink(entity: {type: string, bbid: string}): string {
+	return `/${_.kebabCase(entity.type)}/${entity.bbid}`;
+}
+
+
 export function getNextEnabledAndResultsArray(array, size) {
 	if (array.length > size) {
 		while (array.length > size) {
